requirements.txt
playwright
transformers
accelerate
sentence-transformers
faiss-cpu
requests
nest_asyncio
